@@ -3,7 +3,7 @@ description: Firebase auth flutter
 version: 0.10.0
 publish_to: none
 environment:
-  sdk: '>=2.10.0-0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -13,13 +13,13 @@ dependencies:
     git:
       url: git://github.com/tekartik/firebase_auth.dart
       path: auth
-      ref: dart2
+      ref: null_safety
     version: '>=0.9.0'
   tekartik_firebase_flutter:
     git:
       url: git://github.com/tekartik/firebase_flutter
       path: firebase_flutter
-      ref: dart2
+      ref: null_safety
     version: '>=0.4.0+0'
   firebase_auth: '>=0.18.0+1'
   google_sign_in: '>=4.0.7'
