@@ -1,7 +1,7 @@
 import 'package:tekartik_firebase_auth/auth.dart';
 
 abstract class AuthFlutter {
-  Future<User> googleSignIn();
+  Future<User?> googleSignIn();
 }
 
 abstract class AuthServiceFlutter implements AuthService {}
